+++ ChangeLog.md
@@ -5,6 +5,9 @@
 - optimise validation logical. Ignore readonly and secret property response check
 - DO NOT output log when request method is POST
 - Support level option. Current two levels: request-check, consistency-check. request-check will only check request parameters. consistency-check check both request and response.
+- Improve test scenario markdown report format
+- Bugfix: packaging issue for handlebar files
+- Add test scenario JUnit xml report support
 
 ## 6/10/2021 2.5.5